Optimizing the database interaction with the Submissions.show page
We've replaced the root_comments method with an activerecord relation of
the same name that does the same thing.  This allows us to use
preloading on our root_comments, which is what our view is using as its
starting point to render.

In the controller, when we are reading the submission from the database,
we include a series of .includes() method calls.  Each of these tells
activerecord to preload an additional depth of comments.  We are only
preloading the comment tree four-deep in this example, although if we
were expecting deeply nested conversations, we could go deeper.