Making the submission show page cachable
A specific submission can be stale if that submission model changes, or
if the community it is associated with changes, or if any of its
comments change.

We indicate this to rails, which then constructs an appropriate etag,
with the fresh_when method.

fresh_when and stale? have similar behavior.